Make use of caching update function in more places
