+++ distributed/dashboard/components/nvml.py
@@ -14,8 +14,8 @@ from bokeh.models import (
 from tornado import escape
 from dask.utils import format_bytes
 from distributed.utils import log_errors
-from distributed.dashboard.scheduler import update, BOKEH_THEME
-from distributed.dashboard.utils import without_property_validation
+from distributed.dashboard.scheduler import BOKEH_THEME
+from distributed.dashboard.utils import without_property_validation, update
 
 
 try:
